MVP-26 Update: cg-certificate-registration-screen 이미지 등록 컴포넌트 분리 - custom-image-picker
