+++ app/components/index.ts
@@ -82,3 +82,4 @@ export * from "./cg-calendar/cg-calendar"
 export * from "./go-back-save-next/go-back-save-next"
 export * from "./loading/loading"
 export * from "./registration-notice-note/registration-notice-note"
+export * from "./custom-image-picker/custom-image-picker"
